requirements.txt
discord==2.3.2
python-dotenv==0.20.0
openai==1.3.8
requests==2.26.0
metaphor-python==0.1.23